requirements.txt
streamlit
python-dotenv
openai
requests
PyGithub
jira
langgraph
tiktoken
plotly
networkx
python-docx
openpyxl